Modularized into separate python files based on purpose, and added scripts for running on either Windows or Linux.
