+++ scripts/switch_mac_collector.cmd
@@ -1,0 +1,7 @@
+@echo off
+
+REM Navigate to the src directory
+cd %~dp0\..\src
+
+REM Run the Python script with all arguments passwed to the batch script
+python main.py %*
